// src/app/api/swagger/models/api-trade-dto.ts
export interface ApiTradeDto {
  AdditionalInfo?: string | null;
  AdditionalRequirements?: Array<ApiAdditionalRequirementDto> | null;
  AlwaysSendInvitationsForVendors?: boolean;
  AutoInvitationVendors?: boolean;
  BusinessFlowType?: ApiTradeBusinessFlowEnum;
  CharacteristicFileGuid?: string;
  CharacteristicFromUi?: boolean | null;
  Characteristics?: Array<ApiCharacteristicDto> | null;
  ClientTimeZoneMinutesOffset?: number;
  ConditionsOfPayment?: string | null;
  ContactInfo?: ApiCustomerContactInfoAltDto | null;
  CustomerId?: number;
  CustomerOrgFullName?: string | null;
  CustomerOrganizationId?: number;
  CustomerWorkGroupNames?: Array<string> | null;
  DealDocument?: ApiTradeDocumentDto | null;
  DealExecutionDate?: string | null;
  DeliveryPlace?: string | null;
  DeliveryTerms?: string | null;
  Description?: string | null;
  EnableKtruSpecifications?: boolean;
  EnablePositionCharacteristics?: boolean;
  ExternalContractTemplateId?: number | null;
  ExternalFieldValues?: Array<ApiExternalFieldValueAltDto> | null;
  ExternalId?: string | null;
  ExternalSourcePlatformNumber?: string | null;
  ExternalSystemId?: number | null;
  FederalLawLimit?: ApiFederalLawLimitAltEnum;
  FillingApplicationEndDate?: string | null;
  Id?: number;
  InitialPrice?: number | null;
  InitialPriceWithVat?: number | null;
  InitialPriceWithoutVat?: number | null;
  InvitedOrganizationIds?: Array<number> | null;
  IsAutoCompleteDeal?: boolean;
  IsBasedOnPublicOffer?: boolean | null;
  IsContractInElectronicForm?: boolean | null;
  IsImmediate?: boolean;
  IsInitialPriceDefined?: boolean;
  IsLotPriceWithVat?: boolean;
  IsManualEnterPriceForEachProduct?: boolean;
  IsOnlySmp?: boolean;
  IsParentContractNumberNotSpecified?: boolean | null;
  IsRussianPurchaseObjectHasAdvantage?: boolean;
  IsUnitBidding?: boolean;
  KeepGuaranteeForAllParticipants?: boolean;
  KeepGuaranteeForParticipantsCount?: number | null;
  KladrRegionCode?: string | null;
  KladrRegionCodes?: Array<ApiDictionaryItemAltDto> | null;
  LastModificationDate?: string | null;
  LotItemsSumPrice?: number | null;
  MaxDealPrice?: number | null;
  MinFillingApplicationEndDateForImmediate?: string | null;
  OrganizerId?: number;
  ParentContractNumber?: string | null;
  PlanNumber?: string | null;
  PlanPositionNumber?: string | null;
  PlanedDealSignDate?: string | null;
  Products?: Array<ApiProductDto> | null;
  RegulationDocumentUrl?: string | null;
  ResultsDeliveryType?: number | null;
  RowVersion?: number;
  SourcePlatform?: ApiSourcePlatformEnum;
  SupplierWithSmallBusinessRequirements?: boolean;
  SupplierWithUniversalRequirements?: boolean;
  TradeBiddingType?: ApiTradeBiddingTypeEnum | null;
  TradeDocuments?: Array<ApiTradeDocumentDto> | null;
  TradeFiles?: Array<ApiTradeDocumentDto> | null;
  TradeIdentificationCode?: string | null;
  TradeName?: string | null;
  TradeState?: ApiTradeStateEnum;
  ZmoFinanceSource?: ApiZmoFinanceSourceEnum | null;
  ZmoFzType?: number | null;
}